+++ internal/aghnet/ignore.go
@@ -24,13 +24,13 @@ type IgnoreEngine struct {
 // NewIgnoreEngine creates a new instance of the IgnoreEngine and stores the
 // list of rules for ignoring hostnames.
 func NewIgnoreEngine(ignored []string) (e *IgnoreEngine, err error) {
-	ruleList := []filterlist.Interface{
+	ruleLists := []filterlist.Interface{
 		filterlist.NewString(&filterlist.StringConfig{
 			RulesText:      strings.ToLower(strings.Join(ignored, "\n")),
 			IgnoreCosmetic: true,
 		}),
 	}
-	ruleStorage, err := filterlist.NewRuleStorage(ruleList)
+	ruleStorage, err := filterlist.NewRuleStorage(ruleLists)
 	if err != nil {
 		return nil, err
 	}